Add vs project for examples/webrtc_unity_plugin 🛠
This adds a vs project definition for `webrtc\xplatform\webrtc\examples\unityplugin`, which is required as we need to take a dependency on the wrapperglue for uwp, which does not have any GN targets defined as of yet.

So, in short, to build the `unityplugin` codebase with support for winuwp, you'll need to use the vs project and wrapper layer.