Send ticks to the encoder when dropping frames
When _sampleAttributeQueue is full, the `Encode` method doesn't send the samples
to the SinkWriter but it returns immediately. Somehow this makes the encoder
stall, especially on pipeline initialization - either directly or indirectly by
facilitating some thread scheduling issue.

Sending ticks until the in-flight samples are processed mitigates the issue, by
avoiding the stall under normal load (the stall seems to still happen under
heavy load, e.g. if the pipeline is re-initialized frequently).