Fix invalid H.264 decoder output frame size
When the H.264 decoder returns a decoded frame, the frame might contain
the MF_MT_MINIMUM_DISPLAY_APERTURE attribute. In that case, this
indicates that the frame buffer is padded, and gives the visible area
inside the buffer. Conversely the MF_MT_FRAME_SIZE always gives the
total buffer size. Both are needed to safely access the data in the
buffer and convert it to I420, but the output frame needs to have the
size of the former, the visible area.